+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.1.0] - unreleased
+
+### Added
+
+- Added FreeBusy component, property and period parser/formatter
+- Added Timezone component
+
+### Updated
+
+- Readme with timezone component documentation
+
 ## [1.0.1] - 2020-09-11
 
 - Removed moment.js completely